Display correct accountId on SummaryPage
